@@ -26,15 +26,17 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   aligned_dialog: 0.0.6
+  aligned_tooltip: 0.0.1
   auto_size_text: 3.0.0
   cached_network_image: 3.3.1
   data_table_2: 2.5.10
+  easy_debounce: 2.0.1
   equatable: 2.0.5
   file_picker: ^8.0.7
   flutter_animate: 4.5.0
   flutter_cache_manager: 3.3.1
   flutter_spinkit: 5.2.0
-  font_awesome_flutter: 10.6.0
+  font_awesome_flutter: 10.7.0
   from_css_color: 2.0.0
   go_router: 12.1.3
   google_fonts: 6.1.0
@@ -98,6 +100,38 @@ flutter:
     - assets/lottie_animations/
     - assets/rive_animations/
     - assets/pdfs/
+
+  fonts:
+    - family: 'GoogleSans'
+      fonts:
+        - asset: assets/fonts/GoogleSans-Regular.ttf
+        - asset: assets/fonts/GoogleSans-Italic.ttf
+          style: italic
+        - asset: assets/fonts/GoogleSans-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/GoogleSans-MediumItalic.ttf
+          weight: 500
+          style: italic
+        - asset: assets/fonts/GoogleSans-Bold.ttf
+          weight: 700
+        - asset: assets/fonts/GoogleSans-BoldItalic.ttf
+          weight: 700
+          style: italic
+    - family: 'GoogleSansText'
+      fonts:
+        - asset: assets/fonts/GoogleSansText-Regular.ttf
+        - asset: assets/fonts/GoogleSansText-Italic.ttf
+          style: italic
+        - asset: assets/fonts/GoogleSansText-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/GoogleSansText-MediumItalic.ttf
+          weight: 500
+          style: italic
+        - asset: assets/fonts/GoogleSansText-Bold.ttf
+          weight: 700
+        - asset: assets/fonts/GoogleSansText-BoldItalic.ttf
+          weight: 700
+          style: italic
 
   
   # An image asset can refer to one or more resolution-specific "variants", see
