@@ -11,9 +11,11 @@ class ChatStruct extends BaseStruct {
     String? message,
     bool? isAi,
     String? responseId,
+    bool? isDebugVisible,
   })  : _message = message,
         _isAi = isAi,
-        _responseId = responseId;
+        _responseId = responseId,
+        _isDebugVisible = isDebugVisible;
 
   // "message" field.
   String? _message;
@@ -36,10 +38,18 @@ class ChatStruct extends BaseStruct {
 
   bool hasResponseId() => _responseId != null;
 
+  // "isDebugVisible" field.
+  bool? _isDebugVisible;
+  bool get isDebugVisible => _isDebugVisible ?? true;
+  set isDebugVisible(bool? val) => _isDebugVisible = val;
+
+  bool hasIsDebugVisible() => _isDebugVisible != null;
+
   static ChatStruct fromMap(Map<String, dynamic> data) => ChatStruct(
         message: data['message'] as String?,
         isAi: data['is_ai'] as bool?,
         responseId: data['response_id'] as String?,
+        isDebugVisible: data['isDebugVisible'] as bool?,
       );
 
   static ChatStruct? maybeFromMap(dynamic data) =>
@@ -49,6 +59,7 @@ class ChatStruct extends BaseStruct {
         'message': _message,
         'is_ai': _isAi,
         'response_id': _responseId,
+        'isDebugVisible': _isDebugVisible,
       }.withoutNulls;
 
   @override
@@ -64,6 +75,10 @@ class ChatStruct extends BaseStruct {
         'response_id': serializeParam(
           _responseId,
           ParamType.String,
+        ),
+        'isDebugVisible': serializeParam(
+          _isDebugVisible,
+          ParamType.bool,
         ),
       }.withoutNulls;
 
@@ -84,6 +99,11 @@ class ChatStruct extends BaseStruct {
           ParamType.String,
           false,
         ),
+        isDebugVisible: deserializeParam(
+          data['isDebugVisible'],
+          ParamType.bool,
+          false,
+        ),
       );
 
   @override
@@ -94,20 +114,24 @@ class ChatStruct extends BaseStruct {
     return other is ChatStruct &&
         message == other.message &&
         isAi == other.isAi &&
-        responseId == other.responseId;
+        responseId == other.responseId &&
+        isDebugVisible == other.isDebugVisible;
   }
 
   @override
-  int get hashCode => const ListEquality().hash([message, isAi, responseId]);
+  int get hashCode =>
+      const ListEquality().hash([message, isAi, responseId, isDebugVisible]);
 }
 
 ChatStruct createChatStruct({
   String? message,
   bool? isAi,
   String? responseId,
+  bool? isDebugVisible,
 }) =>
     ChatStruct(
       message: message,
       isAi: isAi,
       responseId: responseId,
+      isDebugVisible: isDebugVisible,
     );
